webservice/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
numpy==1.26.4
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Pillow==10.3.0
scikit-image==0.22.0
httpx==0.25.2